+++ change-log.md
@@ -1,4 +1,9 @@
 - 06/09:
     + set up base project
     + swap react() render plugin to react router plugin, following ReactRouter v7
+    + set up base folders for each feature, implement auth item, auth
+- 07/09:
+    + set up interceptors, dependency injection
+    - state management
+    +